Move QueryExecutor and ServerQueryRequest related classes from pinot-common to pinot-core
Pre-requisite for putting more context into the request to reduce repetitive work on server side
