+++ pinot-core/src/main/java/com/linkedin/pinot/core/query/pruner/ValidSegmentPruner.java
@@ -15,9 +15,9 @@
  */
 package com.linkedin.pinot.core.query.pruner;
 
-import com.linkedin.pinot.common.query.ServerQueryRequest;
 import com.linkedin.pinot.common.segment.SegmentMetadata;
 import com.linkedin.pinot.core.indexsegment.IndexSegment;
+import com.linkedin.pinot.core.query.request.ServerQueryRequest;
 import org.apache.commons.configuration.Configuration;
