+++ pinot-core/src/main/java/com/linkedin/pinot/core/query/request/context/TimerContext.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.linkedin.pinot.common.query.context;
+package com.linkedin.pinot.core.query.request.context;
 
 import com.linkedin.pinot.common.metrics.ServerMetrics;
 import com.linkedin.pinot.common.metrics.ServerQueryPhase;
