+++ pinot-core/src/main/java/com/linkedin/pinot/core/query/scheduler/SchedulerQueryContext.java
@@ -18,7 +18,7 @@ package com.linkedin.pinot.core.query.scheduler;
 import com.google.common.base.Preconditions;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
-import com.linkedin.pinot.common.query.ServerQueryRequest;
+import com.linkedin.pinot.core.query.request.ServerQueryRequest;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
