+++ pinot-core/src/main/java/com/linkedin/pinot/core/query/scheduler/resources/ResourceManager.java
@@ -19,7 +19,7 @@ import com.google.common.annotations.VisibleForTesting;
 import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.common.util.concurrent.MoreExecutors;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
-import com.linkedin.pinot.common.query.ServerQueryRequest;
+import com.linkedin.pinot.core.query.request.ServerQueryRequest;
 import com.linkedin.pinot.core.query.scheduler.SchedulerGroupAccountant;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
