+++ pinot-core/src/main/java/com/linkedin/pinot/core/query/scheduler/resources/UnboundedResourceManager.java
@@ -15,7 +15,7 @@
  */
 package com.linkedin.pinot.core.query.scheduler.resources;
 
-import com.linkedin.pinot.common.query.ServerQueryRequest;
+import com.linkedin.pinot.core.query.request.ServerQueryRequest;
 import com.linkedin.pinot.core.query.scheduler.SchedulerGroupAccountant;
 import org.apache.commons.configuration.Configuration;
