+++ pinot-core/src/test/java/com/linkedin/pinot/core/query/scheduler/resources/ResourceManagerTest.java
@@ -15,7 +15,7 @@
  */
 package com.linkedin.pinot.core.query.scheduler.resources;
 
-import com.linkedin.pinot.common.query.ServerQueryRequest;
+import com.linkedin.pinot.core.query.request.ServerQueryRequest;
 import com.linkedin.pinot.core.query.scheduler.SchedulerGroupAccountant;
 import org.apache.commons.configuration.Configuration;
 import org.apache.commons.configuration.PropertiesConfiguration;
