+++ pinot-server/src/main/java/com/linkedin/pinot/server/starter/ServerInstance.java
@@ -16,8 +16,8 @@
 package com.linkedin.pinot.server.starter;
 
 import com.linkedin.pinot.common.metrics.ServerMetrics;
-import com.linkedin.pinot.common.query.QueryExecutor;
 import com.linkedin.pinot.core.data.manager.InstanceDataManager;
+import com.linkedin.pinot.core.query.executor.QueryExecutor;
 import com.linkedin.pinot.core.query.scheduler.QueryScheduler;
 import com.linkedin.pinot.core.query.scheduler.QuerySchedulerFactory;
 import com.linkedin.pinot.server.conf.ServerConf;
